refactor: 기존 서비스를 추상화하여 헥사고날 적용
