+++ src/main/java/org/example/remedy/domain/song/SongCustomRepository.java
@@ -1,10 +1,8 @@
-package org.example.remedy.application.song;
+package org.example.remedy.domain.song;
 
 import co.elastic.clients.elasticsearch._types.query_dsl.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.example.remedy.domain.song.Song;
-import org.example.remedy.domain.song.SongRepository;
 import org.springframework.data.elasticsearch.client.elc.NativeQuery;
 import org.springframework.data.elasticsearch.core.ElasticsearchOperations;
 import org.springframework.data.elasticsearch.core.SearchHit;
@@ -17,7 +15,7 @@ import java.util.stream.Collectors;
 @Slf4j
 @RequiredArgsConstructor
 @Component
-public class SongQueryBuilder {
+public class SongCustomRepository {
     private final SongRepository songRepository;
     private final ElasticsearchOperations elasticsearchOperations;
